Moved static reflection method in `SimpleEffectDialog` to separate utility class
